# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Seller API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://api.server.test/v3"
    }
  ],
...
  "paths": {
...
    "/store": {
      "get": {
        "summary": "Lists of stores",
        "operationId": "Lists of stores",
        "tags":["store"],
        "responses": {
          "200": {
            "description": "Successfully retrieve the data"
          },
          "400": {
            "description": "You are Not Authorized"
          },
          "408": {
            "description": "Request Timeout"
          }
        }
      },
      "post": {
        "summary": "Add the Store",
        "operationId": "Add the Store",
        "tags":["store"],
        "responses": {
          "201": {
            "description": "Successfully create Store"
          },
          "400": {
            "description": "You are Not Authorized"
          },
          "406": {
            "description": "The form is required, please fill the form"
          },
          "408": {
            "description": "Request Timeout"
          }, 
          "415": {
            "description": "unsuported media type"
          }
        }
      }
    },
...
    "/{storeName}/etalation": {
      "get": {
        "summary": "Get all of etalation from a store name",
        "operationId": "Get all of etalation from a store name",
        "tags":["etalation"],
        "responses": {
          "200": {
            "description": "Successfully retrieve the data"
          },
          "400": {
            "description": "You are Not Authorized"
          },
          "408": {
            "description": "Request Timeout"
          }
        },
        "parameters": [
          {
            "name": "etalation",
            "in": "query",
            "description": "search Etalation in store",
            "required": true,
            "schema": {
              "type": "string",
              "format": "varchar(30)"
            }
          }
        ]
      },
      "post": {
        "summary": "Add the etalation to a store name",
        "operationId": "Add the etalation a store name",
        "tags":["etalation"],
        "responses": {
          "201": {
            "description": "Successfully create user"
          },
          "400": {
            "description": "You are Not Authorized"
          },
          "406": {
            "description": "The form is required, please fill the form"
          },
          "408": {
            "description": "Request Timeout"
          }, 
          "415": {
            "description": "unsuported media type"
          }
        }
      }
    },
...
  },
...
}